Web: Fix broken get_bright_wave precision

+++ src/augs/graphics/rgba.h
@@ -38,7 +38,7 @@ struct rgba {
 
 	static rgba zero;
 
-	static rgba get_bright_wave(const float secs, const float upperLimit1 = 0.58, const float lowerLimit2 = 0.80);
+	static rgba get_bright_wave(const double secs, const double upperLimit1 = 0.58, const double lowerLimit2 = 0.80);
 
 	// GEN INTROSPECTOR struct rgba
 	rgba_channel r;
